2023/CCE/NZK-SIARK/requirements.txt
bcrypt==4.0.1
capstone==5.0.0rc4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colored-traceback==0.3.0
cryptography==41.0.1
idna==3.4
intervaltree==3.1.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
paramiko==3.2.0
plumbum==1.8.2
psutil==5.9.5
pwntools==4.10.0
pycparser==2.21
pyelftools==0.29
Pygments==2.15.1
PyNaCl==1.5.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
ROPGadget==7.3
rpyc==5.3.1
six==1.16.0
sortedcontainers==2.4.0
unicorn==2.0.1.post1
urllib3==2.0.3
